Divide Python target into submodules
[changelog skip]
